+++ Sources/KeyboardKit/System/SystemKeyboardButtonRowItem.swift
@@ -69,7 +69,7 @@ public struct SystemKeyboardButtonRowItem<Content: View>: View {
             .frame(maxWidth: .infinity)
             .frame(height: item.size.height - item.insets.top - item.insets.bottom)
             .rowItemWidth(for: item, totalWidth: keyboardWidth, referenceWidth: inputWidth)
-            .systemKeyboardButtonStyle(
+            .keyboardButtonStyle(
                 buttonStyle,
                 isPressed: isPressed
             )